+++ src/background/modules/cli/mod.rs
@@ -3,43 +3,44 @@ pub mod domain;
 
 use std::{
     fs,
-    io::{BufReader, Read},
     net::{TcpListener, TcpStream},
+    path::PathBuf,
 };
 
 use application::{handle_cli_events, SEELEN_COMMAND_LINE};
+use itertools::Itertools;
+use tauri_plugin_dialog::{DialogExt, MessageDialogKind};
 
 use crate::{
-    error_handler::Result, log_error, seelen::Seelen, trace_lock, utils::spawn_named_thread,
+    error_handler::Result,
+    log_error,
+    seelen::{get_app_handle, Seelen},
+    trace_lock,
+    utils::{pwsh::PwshScript, spawn_named_thread},
 };
 
-pub struct Client;
-impl Client {
-    // const BUFFER_SIZE: usize = 5 * 1024 * 1024; // 5 MB
+pub struct AppClient;
+impl AppClient {
+    fn socket_path() -> PathBuf {
+        std::env::temp_dir().join("com.seelen.seelen-ui\\slu_tcp_socket")
+    }
 
+    // const BUFFER_SIZE: usize = 5 * 1024 * 1024; // 5 MB
     fn handle_message(stream: TcpStream) {
-        let mut reader = BufReader::new(stream);
-        let mut buffer = vec![];
-        match reader.read_to_end(&mut buffer) {
-            Ok(_) => {
-                let message = String::from_utf8_lossy(&buffer).to_string();
-                match serde_json::from_str::<Vec<String>>(&message) {
-                    Ok(argv) => {
-                        log::trace!(target: "slu::cli", "{}", argv[1..].join(" "));
-                        std::thread::spawn(move || {
-                            let command = trace_lock!(SEELEN_COMMAND_LINE).clone();
-                            log_error!(handle_cli_events(&command.get_matches_from(argv)));
-                        });
-                    }
-                    Err(e) => {
-                        log::error!("Failed to deserialize message: {}", e);
-                    }
-                }
-            }
+        let argv = match serde_json::from_reader::<TcpStream, Vec<String>>(stream) {
+            Ok(argv) => argv,
             Err(e) => {
-                log::error!("Failed to read from stream: {}", e);
+                log::error!("Failed to deserialize message: {}", e);
+                return;
             }
-        }
+        };
+        log::trace!(target: "slu::cli", "{}", argv[1..].join(" "));
+        std::thread::spawn(move || {
+            let command = trace_lock!(SEELEN_COMMAND_LINE).clone();
+            if let Ok(matches) = command.try_get_matches_from(argv) {
+                log_error!(handle_cli_events(&matches));
+            }
+        });
     }
 
     pub fn listen_tcp() -> Result<()> {
@@ -48,10 +49,7 @@ impl Client {
         let port = socket_addr.port();
 
         log::info!("TCP server listening on 127.0.0.1:{}", port);
-        fs::write(
-            std::env::temp_dir().join("slu_tcp_socket"),
-            port.to_string(),
-        )?;
+        fs::write(Self::socket_path(), port.to_string())?;
 
         spawn_named_thread("TCP Listener", move || {
             for stream in listener.incoming() {
@@ -61,9 +59,7 @@ impl Client {
                 }
                 match stream {
                     Ok(stream) => Self::handle_message(stream),
-                    Err(e) => {
-                        log::error!("Failed to accept connection: {}", e);
-                    }
+                    Err(e) => log::error!("Failed to accept connection: {}", e),
                 }
             }
         })?;
@@ -71,7 +67,79 @@ impl Client {
     }
 
     pub fn connect_tcp() -> Result<TcpStream> {
-        let port = fs::read_to_string(std::env::temp_dir().join("slu_tcp_socket"))?;
+        let port = fs::read_to_string(Self::socket_path())?;
         Ok(TcpStream::connect(format!("127.0.0.1:{}", port))?)
+    }
+
+    pub fn redirect_cli_to_instance() -> Result<()> {
+        let mut attempts: i32 = 0;
+        let mut stream = Self::connect_tcp();
+        while stream.is_err() && attempts < 10 {
+            attempts += 1;
+            std::thread::sleep(std::time::Duration::from_millis(100));
+            stream = AppClient::connect_tcp();
+        }
+        serde_json::to_writer(stream?, &std::env::args().collect_vec())?;
+        Ok(())
+    }
+}
+
+pub struct ServiceClient;
+impl ServiceClient {
+    fn token() -> &'static str {
+        std::option_env!("SLU_SERVICE_CONNECTION_TOKEN").unwrap_or("__local__")
+    }
+
+    fn socket_path() -> PathBuf {
+        PathBuf::from(r"C:\Windows\Temp\slu_service_tcp_socket")
+    }
+
+    fn connect_tcp() -> Result<TcpStream> {
+        let port = fs::read_to_string(Self::socket_path())?;
+        Ok(TcpStream::connect(format!("127.0.0.1:{}", port))?)
+    }
+
+    fn send_message(message: &[&str]) -> Result<()> {
+        let stream = Self::connect_tcp()?;
+        let writter = std::io::BufWriter::new(stream);
+        serde_json::to_writer(
+            writter,
+            &serde_json::json!({
+                "token": Self::token(),
+                "message": message,
+            }),
+        )?;
+        Ok(())
+    }
+
+    pub fn is_running() -> bool {
+        Self::connect_tcp().is_ok()
+    }
+
+    // the service should be running since installer will start it or startup task scheduler
+    // so if the service is not running, we need to reinstall it (common on msix setup)
+    pub async fn reinstall_and_start() -> Result<()> {
+        get_app_handle()
+            .dialog()
+            .message(t!("service.not_running_description"))
+            .title(t!("service.not_running"))
+            .kind(MessageDialogKind::Info)
+            .ok_button_label(t!("service.not_running_ok"))
+            .blocking_show();
+
+        log::info!("Reinstalling service...");
+        let service_path = std::env::current_exe()?.with_file_name("slu-service.exe");
+        PwshScript::new(format!(
+            "&\"{}\" install \n net start slu-service",
+            service_path.display(),
+        ))
+        .elevated()
+        .execute()
+        .await?;
+        Ok(())
+    }
+
+    pub fn emit_stop_signal() -> Result<()> {
+        Self::send_message(&["stop"])
     }
 }