+++ src/background/seelen_weg/weg_items_impl.rs
@@ -9,6 +9,7 @@ use tauri::Emitter;
 
 use crate::{
     error_handler::Result,
+    log_error,
     modules::start::application::START_MENU_MANAGER,
     seelen::get_app_handle,
     state::application::FULL_STATE,
@@ -124,7 +125,7 @@ impl WegItemsImpl {
 
         let get_info_from = match &umid {
             Some(umid) => {
-                let _ = extract_and_save_icon_umid(umid);
+                log_error!(extract_and_save_icon_umid(umid));
                 if WindowsApi::is_uwp_package_id(umid) {
                     ShouldGetInfoFrom::Package(umid.clone())
                 } else {
@@ -134,6 +135,7 @@ impl WegItemsImpl {
             None => ShouldGetInfoFrom::Process,
         };
 
+        let mut pin_disabled = false;
         let mut display_name = window
             .app_display_name()
             .unwrap_or_else(|_| String::from("Unknown"));
@@ -159,7 +161,7 @@ impl WegItemsImpl {
                         .to_string_lossy()
                         .to_string();
                 } else {
-                    let _ = extract_and_save_icon_from_file(&path);
+                    log_error!(extract_and_save_icon_from_file(&path));
                 }
                 // System.AppUserModel.RelaunchCommand and System.AppUserModel.RelaunchDisplayNameResource
                 // must always be set together. If one of those properties is not set, then neither is used.
@@ -170,11 +172,12 @@ impl WegItemsImpl {
                     relaunch_command = win_relaunch_command;
                     display_name = relaunch_display_name;
                 } else {
+                    pin_disabled = window.prevent_pinning();
                     relaunch_command = format!("\"explorer.exe\" shell:AppsFolder\\{umid}");
                 }
             }
             ShouldGetInfoFrom::Process => {
-                let _ = extract_and_save_icon_from_file(&path);
+                log_error!(extract_and_save_icon_from_file(&path));
                 relaunch_command = path.to_string_lossy().to_string();
             }
         };
@@ -216,6 +219,7 @@ impl WegItemsImpl {
                 title: window.title(),
                 handle: window.address(),
             }],
+            pin_disabled,
         };
         self.items.center.push(WegItem::Temporal(data));
         Ok(())